+++ packages/terra-menu/src/MenuItem.jsx
@@ -22,6 +22,11 @@ const propTypes = {
   text: PropTypes.string,
 
   /**
+   * Indicates if item should be disabled
+   */
+  isDisabled: PropTypes.bool,
+
+  /**
    * Indicates if the item is selected. A selected item is indicated with a checkmark.
    */
   isSelected: PropTypes.bool,
@@ -57,6 +62,7 @@ const defaultProps = {
   isSelected: false,
   isActive: false,
   isSelectable: undefined,
+  isDisabled: false,
   subMenuItems: [],
 };
 
@@ -118,6 +124,7 @@ class MenuItem extends React.Component {
   render() {
     const {
       text,
+      isDisabled,
       isSelected,
       isSelectable,
       subMenuItems,
@@ -128,11 +135,16 @@ class MenuItem extends React.Component {
     const { isGroupItem, isSelectableMenu } = this.context;
 
     const attributes = Object.assign({}, customProps);
-    attributes.tabIndex = '0';
+    attributes.tabIndex = isDisabled ? '-1' : '0';
+    attributes['aria-disabled'] = isDisabled;
+
     // This is passed down by the single select list in group item and not needed
     delete attributes.hasChevron;
 
-    if (isSelectable && !isGroupItem) {
+    if (isDisabled) {
+      delete attributes.onClick;
+      delete attributes.onKeyDown;
+    } else if (isSelectable && !isGroupItem && !isDisabled) {
       attributes.onClick = this.wrapOnClick;
       attributes.onKeyDown = this.wrapOnKeyDown(attributes.onKeyDown);
     }
@@ -140,6 +152,7 @@ class MenuItem extends React.Component {
     const itemClassNames = cx([
       'item',
       { selected: this.state.isSelected || (isGroupItem && isSelected) },
+      { 'is-disabled': isDisabled },
       attributes.className,
     ]);