+++ packages/terra-site/src/examples/menu/BasicMenu.jsx
@@ -110,6 +110,13 @@ class BasicMenu extends React.Component {
               onClick={this.handleToggle2OnClick}
               isSelectable
             />
+            <Menu.Item
+              text="Disabled Item 1"
+              key="Disabled1"
+              isSelected={this.state.toggle1Selected}
+              onClick={this.handleToggle1OnClick}
+              isDisabled
+            />
             <Menu.Divider key="Divider1" />
             <Menu.Item
               text="Nested Menu 1"
@@ -127,6 +134,7 @@ class BasicMenu extends React.Component {
             <Menu.Item
               text="Nested Menu 2 has a long title that will wrap and a truncated title when clicked"
               key="Nested2"
+              isDisabled
               subMenuItems={[
                 <Menu.Item text="Default 2.1" key="2.1" />,
                 <Menu.Item text="Default 2.2" key="2.2" />,
@@ -140,6 +148,7 @@ class BasicMenu extends React.Component {
             <Menu.ItemGroup key="Group" onChange={this.handleOnChange}>
               <Menu.Item text="Group Item 1" key="GroupItem1" isSelected={this.state.groupSelectedIndex === 0} />
               <Menu.Item text="Group Item 2" key="GroupItem2" isSelected={this.state.groupSelectedIndex === 1} />
+              <Menu.Item text="Group Item 3" key="GroupItem3" isSelected={this.state.groupSelectedIndex === 2} isDisabled />
             </Menu.ItemGroup>
           </Menu>
           <Button onClick={this.handleButtonClick} text="Click Me" />