+++ packages/terra-menu/CHANGELOG.md
@@ -4,7 +4,8 @@ Changelog
 Unreleased
 ----------
 ### Added
-Keyboard navigation with arrow keys
+* Keyboard navigation with arrow keys
+* Added ability to disable menu items
 
 1.2.0 - (September 26, 2017)
 ------------------