+++ packages/terra-menu/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+#Added
+* Added ability to disable menu items
 
 1.2.0 - (September 26, 2017)
 ------------------