+++ packages/terra-menu/CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 Unreleased
 ----------
 
+1.3.0 - (October 6, 2017)
+------------------
 ### Fixed
 * Bug around selectable menus with only 1 child element.